Revert "remove wrong warning"
Re-introduce a check for gamma and lambda fixed to different values in
case of equal asymptote experiments, but turning it into an Exception.
There's no way in this case to know what the user meant, so it safer
to bail out instead of guessing which parameter is to be trusted.

This reverts commit ec4e71629b9974f90343a2fcd7ae8ab0046681ce.