converted all the How-to pages to .md format and made sure all the files get build in the html version
I have not yet changed the content of the "how to" files, only converted the fileformat from rst to md.
When building the html docs, I encounterd the bug, that the How-to files did not get build, so i found out that there were some typos in the path in the _toc.yml file and fixed them. After that there was still a warning when building the html docs, because the parameter_recovery_demo file (from examples) was not included in the _toc.yml file and therefore didn't get build. This actually led to the how-to-change-threshold-percent file to be shown as "1. Basic usage".  When I included the parameter recovery demo file in _toc.yml, all the how-to files were build correctly.
I will now start to change and format their content.