re-instate the Result.parameter_estimate attribute
The attribute is now a property, where the getter is Result.get_parameter_estimate, so that in our
private code and for plotting we can still use the method Result.get_parameter_estimate.

For backward compatibility and for simmetry naive users can just access the attribute
Result.parameter_estimate which will return the parameters estimated with the default
estimation method, i.e. 'MAP'.