get rid of the doc page with the MATLAB comparison
The data and the plots showed here are obsolete. The rafactored
implementation of python-psignifit does not reproduce the results
of the MATLAB implementation down to numerical precision because
some different choices have been made. Nothing really relevant, but
the numbers do not match exactly anymore. Also, for the user of the
documentation the comparison with the MATLAB version is not interesting.
The MATLAB implementation is much less relevant than the battery of
tests we have in python-psignifit if you care about precision.