add cast_np_scalar utility to cast numpy scalars to Python int/floats
As a result of the implementation of NEP 51
https://numpy.org/neps/nep-0051-scalar-representation.html
numpy scalars have a repr of the form np.float64(1.2), i.e. they do not look like Python floats
anymore. This is not a numerical problem, but it is indeed confusing for the user, which may not
be familiar with numpy scalars. This function converts numpy scalars to Python scalars and let's
everything else (notably ndarrays) untouched, so it can be used unconditionally on anything we
return and expect to be a Python scalar.